requirements.txt
Flask==3.0.1
gunicorn==22.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31

# Optional but commonly used packages
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.10.0
Flask-CORS==4.0.0

python-dotenv
pytest
selenium==4.11.2
urllib3==2.0.4
